+++ tests/src/rules/no-nodejs-modules.js
@@ -26,6 +26,36 @@ ruleTester.run('no-nodejs-modules', rule, {
     test({ code: 'var foo = require("foo")'}),
     test({ code: 'var foo = require("./")'}),
     test({ code: 'var foo = require("@scope/foo")'}),
+    test({
+      code: 'import events from "events"',
+      options: [{
+        allow: ['events'],
+      }],
+    }),
+    test({
+      code: 'import path from "path"',
+      options: [{
+        allow: ['path'],
+      }],
+    }),
+    test({
+      code: 'var events = require("events")',
+      options: [{
+        allow: ['events'],
+      }],
+    }),
+    test({
+      code: 'var path = require("path")',
+      options: [{
+        allow: ['path'],
+      }],
+    }),
+    test({
+      code: 'import path from "path";import events from "events"',
+      options: [{
+        allow: ['path', 'events'],
+      }],
+    }),
   ],
   invalid: [
     test({
@@ -42,6 +72,13 @@ ruleTester.run('no-nodejs-modules', rule, {
     }),
     test({
       code: 'var fs = require("fs")',
+      errors: [error('Do not import Node.js builtin module "fs"')],
+    }),
+    test({
+      code: 'import fs from "fs"',
+      options: [{
+        allow: ['path'],
+      }],
       errors: [error('Do not import Node.js builtin module "fs"')],
     }),
   ],